requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
fake-useragent==1.3.0
flake8==6.1.0
flake8-commas==2.1.0
flake8-docstrings==1.7.0
flake8-print==5.0.0
flake8-quotes==3.3.2
h11==0.14.0
idna==3.4
importlib-resources==6.1.0
inflection==0.5.1
isort==5.12.0
lxml==4.9.3
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycodestyle==2.11.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.1.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.6
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
zipp==3.17.0
